np.float to float

+++ trulens_eval/trulens_eval/tru_db.py
@@ -5,28 +5,27 @@ import logging
 from pathlib import Path
 from pprint import PrettyPrinter
 import sqlite3
-from typing import (Dict, Iterable, List, Optional, Sequence, Tuple)
-
-import numpy as np
+from typing import Dict, Iterable, List, Optional, Sequence, Tuple
 
 from frozendict import frozendict
 from merkle_json import MerkleJson
+import numpy as np
 import pandas as pd
 
 from trulens_eval.schema import ChainID
+from trulens_eval.schema import Cost
 from trulens_eval.schema import FeedbackDefinition
 from trulens_eval.schema import FeedbackDefinitionID
 from trulens_eval.schema import FeedbackResult
 from trulens_eval.schema import FeedbackResultID
+from trulens_eval.schema import FeedbackResultStatus
 from trulens_eval.schema import JSONPath
 from trulens_eval.schema import Model
 from trulens_eval.schema import Record
 from trulens_eval.schema import RecordChainCall
 from trulens_eval.schema import RecordID
-from trulens_eval.schema import Cost
-from trulens_eval.schema import FeedbackResultStatus
-from trulens_eval.util import GetItemOrAttribute
 from trulens_eval.util import all_queries
+from trulens_eval.util import GetItemOrAttribute
 from trulens_eval.util import JSON
 from trulens_eval.util import json_str_of_obj
 from trulens_eval.util import JSONPath
@@ -591,7 +590,7 @@ class LocalSQLite(TruDB):
         df_results = df_results.drop(columns=["name", "result", "calls_json"])
 
         def nonempty(val):
-            if isinstance(val, np.float):
+            if isinstance(val, float):
                 return not np.isnan(val)
             return True
